+++ lib/irb/history.rb
@@ -1,55 +1,4 @@
-# frozen_string_literal: false
-#
-#   save-history.rb -
-#   	by Keiju ISHITSUKA(keiju@ruby-lang.org)
-#
-
 module IRB
-  module HistorySavingAbility # :nodoc:
-  end
-
-  class Context
-    def init_save_history# :nodoc:
-      unless (class<<@io;self;end).include?(HistorySavingAbility)
-        @io.extend(HistorySavingAbility)
-      end
-    end
-
-    # A copy of the default <code>IRB.conf[:SAVE_HISTORY]</code>
-    def save_history
-      IRB.conf[:SAVE_HISTORY]
-    end
-
-    remove_method(:save_history=) if method_defined?(:save_history=)
-    # Sets <code>IRB.conf[:SAVE_HISTORY]</code> to the given +val+ and calls
-    # #init_save_history with this context.
-    #
-    # Will store the number of +val+ entries of history in the #history_file
-    #
-    # Add the following to your +.irbrc+ to change the number of history
-    # entries stored to 1000:
-    #
-    #     IRB.conf[:SAVE_HISTORY] = 1000
-    def save_history=(val)
-      IRB.conf[:SAVE_HISTORY] = val
-      if val
-        main_context = IRB.conf[:MAIN_CONTEXT]
-        main_context = self unless main_context
-        main_context.init_save_history
-      end
-    end
-
-    # A copy of the default <code>IRB.conf[:HISTORY_FILE]</code>
-    def history_file
-      IRB.conf[:HISTORY_FILE]
-    end
-
-    # Set <code>IRB.conf[:HISTORY_FILE]</code> to the given +hist+.
-    def history_file=(hist)
-      IRB.conf[:HISTORY_FILE] = hist
-    end
-  end
-
   module HistorySavingAbility # :nodoc:
     def HistorySavingAbility.extended(obj)
       IRB.conf[:AT_EXIT].push proc{obj.save_history}
